Rename files, split Task and Tasks
In a subsequent commit, the global variables containing state should become
enclosed in a suitable closure.  In that situation the modules would export
factory functions, which makes upper case letters more appropriate.
